+++ src/ExLibs/session.py
@@ -1,5 +1,5 @@
 import time as t
-from feeder import  gpio_feed
+
 import numpy as np
 from collections import deque
 import pandas as pd
@@ -7,7 +7,9 @@ from datetime import datetime
 from datetime import timedelta
 import RPi.GPIO as GPIO
 
-from trial import Trial
+from ExLibs.trial import Trial
+from ExLibs.feeder import  gpio_feed
+
 
 STATE_IDLE = 0
 STATE_TRIAL_INIT = 1 #probably not necessary
